requirements.txt
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
h11==0.14.0
idna==3.10
numpy==2.2.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pycparser==2.22
pyotp==2.9.0
PySocks==1.7.1
pytesseract==0.3.13
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0